+++ fud2/src/exec/mod.rs
@@ -1,0 +1,8 @@
+mod data;
+mod driver;
+mod request;
+
+pub use data::{OpRef, SetupRef, StateRef};
+pub(super) use data::{Operation, Setup, State};
+pub use driver::{Driver, DriverBuilder, Plan};
+pub use request::Request;
