+++ fud2/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod cli;
 pub mod config;
-pub mod driver;
+pub mod exec;
 pub mod run;
+pub mod utils;
 
-pub use driver::{Driver, DriverBuilder};
+pub use exec::{Driver, DriverBuilder};
